[GUI]: window appears cut for some dual-monitor users
Hardcoding the window position is causing a cut window for anyone who has a small monitor on the left side of a larger one, see "https://i.imgur.com/nvtjyOx.png" for an example.

Will fix it with a mix between letting the operating system handle the window position and your desired offset.